Enemy at bottom fix
Enemies weren't removed when they reach bottom of the screen. Instead they were removed when they are over the top of it (which doesn't happen in the gameplay). It caused all enemies to stay in memory even after they flew away.
This fix makes enemies to be removed when the reach the bottom.